Dashboard
Dashboard, containing database client code handling the stuff originally on node-red. Everything containerized within a docker container. A few things left to add for the mqtt such as last will message, handling of the deviceStatus, displaying last ping of the device to show whether a device is connected or not; and userAccess to see what user accessed what gate and when